Update: Gate for update event and remove attendee

+++ app/Http/Controllers/Api/AttendeeController.php
@@ -2,18 +2,24 @@
 
 namespace App\Http\Controllers\Api;
 
-use App\Http\Controllers\Controller;
+
 use App\Http\Resources\AttendeeResource;
 use App\Http\Traits\CanLoadRelationship;
 use App\Models\Attendee;
 use App\Models\Event;
 use Illuminate\Http\Request;
+use Illuminate\Routing\Controller;
+use Illuminate\Support\Facades\Gate;
 
 class AttendeeController extends Controller
 {
     use CanLoadRelationship;
 
     private array $relations = ['user'];
+    public function __construct()
+    {
+        $this->middleware('auth:sanctum')->except(['index', 'show', 'update']);
+    }
     /**
      * Display a listing of the resource.
      */
@@ -34,7 +40,7 @@ class AttendeeController extends Controller
     public function store(Request $request, Event $event)
     {
         $attendee = $this->loadRelationship($event->attendees()->create([
-            'user_id' => 1,
+            'user_id' => $request->user()->id,
         ])
     );
 
@@ -54,8 +60,12 @@ class AttendeeController extends Controller
     /**
      * Remove the specified resource from storage.
      */
-    public function destroy(string $event, Attendee $attendee)
+    public function destroy(Event $event, Attendee $attendee)
     {
+        if(Gate::denies('delete-attendee', [$event, $attendee])){
+            abort(403,'You are not allowed to delete this attendee');
+        }
+
         $attendee->delete();
         return response(status: 204);
     }
